// FRONTEND/chatbot-admin-panel/src/services/api/schemas/contextDefinitionResponse.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { ContextDefinitionResponseDescription } from './contextDefinitionResponseDescription';
import type { ContextMainType } from './contextMainType';
import type { ContextDefinitionResponseDefaultLlmModelConfigId } from './contextDefinitionResponseDefaultLlmModelConfigId';
import type { ContextDefinitionResponseVirtualAgentProfileId } from './contextDefinitionResponseVirtualAgentProfileId';
import type { DocumentSourceResponse } from './documentSourceResponse';
import type { ContextDefinitionResponseDbConnectionConfig } from './contextDefinitionResponseDbConnectionConfig';
import type { ContextDefinitionResponseDefaultLlmModelConfig } from './contextDefinitionResponseDefaultLlmModelConfig';
import type { ContextDefinitionResponseVirtualAgentProfile } from './contextDefinitionResponseVirtualAgentProfile';
import type { ContextDefinitionResponseProcessingConfigDocumental } from './contextDefinitionResponseProcessingConfigDocumental';
import type { ContextDefinitionResponseProcessingConfigDatabaseQuery } from './contextDefinitionResponseProcessingConfigDatabaseQuery';
import type { ContextDefinitionResponseProcessingConfigRaw } from './contextDefinitionResponseProcessingConfigRaw';

export interface ContextDefinitionResponse {
  /**
   * @minLength 3
   * @maxLength 150
   */
  name: string;
  description?: ContextDefinitionResponseDescription;
  is_active?: boolean;
  is_public?: boolean;
  main_type: ContextMainType;
  default_llm_model_config_id?: ContextDefinitionResponseDefaultLlmModelConfigId;
  virtual_agent_profile_id?: ContextDefinitionResponseVirtualAgentProfileId;
  id: number;
  created_at: string;
  updated_at: string;
  document_sources?: DocumentSourceResponse[];
  db_connection_config?: ContextDefinitionResponseDbConnectionConfig;
  default_llm_model_config?: ContextDefinitionResponseDefaultLlmModelConfig;
  virtual_agent_profile?: ContextDefinitionResponseVirtualAgentProfile;
  processing_config_documental?: ContextDefinitionResponseProcessingConfigDocumental;
  processing_config_database_query?: ContextDefinitionResponseProcessingConfigDatabaseQuery;
  processing_config_raw?: ContextDefinitionResponseProcessingConfigRaw;
}