// FRONTEND/chatbot-admin-panel/src/services/api/schemas/contextDefinitionCreate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { ContextDefinitionCreateDescription } from './contextDefinitionCreateDescription';
import type { ContextMainType } from './contextMainType';
import type { ContextDefinitionCreateDefaultLlmModelConfigId } from './contextDefinitionCreateDefaultLlmModelConfigId';
import type { ContextDefinitionCreateVirtualAgentProfileId } from './contextDefinitionCreateVirtualAgentProfileId';
import type { ContextDefinitionCreateDbConnectionConfigId } from './contextDefinitionCreateDbConnectionConfigId';
import type { ContextDefinitionCreateProcessingConfigDocumental } from './contextDefinitionCreateProcessingConfigDocumental';
import type { ContextDefinitionCreateProcessingConfigDatabaseQuery } from './contextDefinitionCreateProcessingConfigDatabaseQuery';

export interface ContextDefinitionCreate {
  /**
   * @minLength 3
   * @maxLength 150
   */
  name: string;
  description?: ContextDefinitionCreateDescription;
  is_active?: boolean;
  is_public?: boolean;
  main_type: ContextMainType;
  default_llm_model_config_id?: ContextDefinitionCreateDefaultLlmModelConfigId;
  virtual_agent_profile_id?: ContextDefinitionCreateVirtualAgentProfileId;
  document_source_ids?: number[];
  db_connection_config_id?: ContextDefinitionCreateDbConnectionConfigId;
  processing_config_documental?: ContextDefinitionCreateProcessingConfigDocumental;
  processing_config_database_query?: ContextDefinitionCreateProcessingConfigDatabaseQuery;
}