// FRONTEND/chatbot-admin-panel/src/services/api/schemas/contextDefinitionUpdate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { ContextDefinitionUpdateName } from './contextDefinitionUpdateName';
import type { ContextDefinitionUpdateDescription } from './contextDefinitionUpdateDescription';
import type { ContextDefinitionUpdateIsActive } from './contextDefinitionUpdateIsActive';
import type { ContextDefinitionUpdateIsPublic } from './contextDefinitionUpdateIsPublic';
import type { ContextDefinitionUpdateMainType } from './contextDefinitionUpdateMainType';
import type { ContextDefinitionUpdateDefaultLlmModelConfigId } from './contextDefinitionUpdateDefaultLlmModelConfigId';
import type { ContextDefinitionUpdateVirtualAgentProfileId } from './contextDefinitionUpdateVirtualAgentProfileId';
import type { ContextDefinitionUpdateDocumentSourceIds } from './contextDefinitionUpdateDocumentSourceIds';
import type { ContextDefinitionUpdateDbConnectionConfigId } from './contextDefinitionUpdateDbConnectionConfigId';
import type { ContextDefinitionUpdateProcessingConfigDocumental } from './contextDefinitionUpdateProcessingConfigDocumental';
import type { ContextDefinitionUpdateProcessingConfigDatabaseQuery } from './contextDefinitionUpdateProcessingConfigDatabaseQuery';

export interface ContextDefinitionUpdate {
  name?: ContextDefinitionUpdateName;
  description?: ContextDefinitionUpdateDescription;
  is_active?: ContextDefinitionUpdateIsActive;
  is_public?: ContextDefinitionUpdateIsPublic;
  main_type?: ContextDefinitionUpdateMainType;
  default_llm_model_config_id?: ContextDefinitionUpdateDefaultLlmModelConfigId;
  virtual_agent_profile_id?: ContextDefinitionUpdateVirtualAgentProfileId;
  document_source_ids?: ContextDefinitionUpdateDocumentSourceIds;
  db_connection_config_id?: ContextDefinitionUpdateDbConnectionConfigId;
  processing_config_documental?: ContextDefinitionUpdateProcessingConfigDocumental;
  processing_config_database_query?: ContextDefinitionUpdateProcessingConfigDatabaseQuery;
}