// FRONTEND/chatbot-admin-panel/src/services/api/schemas/hTTPValidationError.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { ValidationError } from './validationError';

export interface HTTPValidationError {
  detail?: ValidationError[];
}